+++ package.json
@@ -11,7 +11,7 @@
     "url": "bendrucker.me"
   },
   "scripts": {
-    "test": "standard && tape test.js"
+    "test": "standard && tape test/*.js"
   },
   "keywords": [
     "functional",
